Add new homepage, moved orignal homepage to blog.js, created Nav comp.
