+++ docs/page-data/using-typescript/page-data.json
@@ -1,1 +1,1 @@
-{"componentChunkName":"component---src-pages-using-typescript-tsx","path":"/using-typescript/","result":{"data":{"site":{"buildTime":"2021-04-05 05:53 pm UTC"}},"pageContext":{}},"staticQueryHashes":["2841359383"]}
+{"componentChunkName":"component---src-pages-using-typescript-tsx","path":"/using-typescript/","result":{"data":{"site":{"buildTime":"2021-04-07 02:27 am UTC"}},"pageContext":{}},"staticQueryHashes":["2841359383"]}
