+++ docs/rss.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="UTF-8"?><rss xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:content="http://purl.org/rss/1.0/modules/content/" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0"><channel><title><![CDATA[Kenny Lee's personal blog]]></title><description><![CDATA[Kenny's personal thoughts.]]></description><link>https://kennylee.com</link><generator>GatsbyJS</generator><lastBuildDate>Mon, 05 Apr 2021 17:54:08 GMT</lastBuildDate><item><title><![CDATA[Learning React & Gatsby]]></title><description><![CDATA[I am relearning how to use React and Gatsby, starting with this blog.]]></description><link>https://kennylee.com/2020-04-01-learning-react-gatsby/</link><guid isPermaLink="false">https://kennylee.com/2020-04-01-learning-react-gatsby/</guid><pubDate>Thu, 01 Apr 2021 00:00:00 GMT</pubDate><content:encoded>&lt;p&gt;I am relearning how to use React and Gatsby, starting with this blog.&lt;/p&gt;</content:encoded></item><item><title><![CDATA[href attribute 'do nothing' trick]]></title><description><![CDATA[I learned this trick from Chris Coyier was speaking at Event Apart Seattle 2017. It’s a common practice to set the anchor element’s href to…]]></description><link>https://kennylee.com/2017-04-10-href-attribute-trick/</link><guid isPermaLink="false">https://kennylee.com/2017-04-10-href-attribute-trick/</guid><pubDate>Mon, 10 Apr 2017 00:13:21 GMT</pubDate><content:encoded>&lt;p&gt;I learned this trick from Chris Coyier was speaking at &lt;a href=&quot;https://aneventapart.com/event/seattle-2017#chris-coyier&quot;&gt;Event Apart Seattle 2017&lt;/a&gt;.&lt;/p&gt;
+<?xml version="1.0" encoding="UTF-8"?><rss xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:content="http://purl.org/rss/1.0/modules/content/" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0"><channel><title><![CDATA[Kenny Lee's personal blog]]></title><description><![CDATA[Kenny's personal thoughts.]]></description><link>https://kennylee.com</link><generator>GatsbyJS</generator><lastBuildDate>Wed, 07 Apr 2021 02:27:40 GMT</lastBuildDate><item><title><![CDATA[Learning React & Gatsby]]></title><description><![CDATA[I am relearning how to use React and Gatsby, starting with this blog.]]></description><link>https://kennylee.com/2020-04-01-learning-react-gatsby/</link><guid isPermaLink="false">https://kennylee.com/2020-04-01-learning-react-gatsby/</guid><pubDate>Thu, 01 Apr 2021 00:00:00 GMT</pubDate><content:encoded>&lt;p&gt;I am relearning how to use React and Gatsby, starting with this blog.&lt;/p&gt;</content:encoded></item><item><title><![CDATA[href attribute 'do nothing' trick]]></title><description><![CDATA[I learned this trick from Chris Coyier was speaking at Event Apart Seattle 2017. It’s a common practice to set the anchor element’s href to…]]></description><link>https://kennylee.com/2017-04-10-href-attribute-trick/</link><guid isPermaLink="false">https://kennylee.com/2017-04-10-href-attribute-trick/</guid><pubDate>Mon, 10 Apr 2017 00:13:21 GMT</pubDate><content:encoded>&lt;p&gt;I learned this trick from Chris Coyier was speaking at &lt;a href=&quot;https://aneventapart.com/event/seattle-2017#chris-coyier&quot;&gt;Event Apart Seattle 2017&lt;/a&gt;.&lt;/p&gt;
 &lt;p&gt;It’s a common practice to set the anchor element’s href to &lt;code class=&quot;language-text&quot;&gt;#&lt;/code&gt;, when you don’t want anything to happen when user click on the link.&lt;/p&gt;
 &lt;div class=&quot;gatsby-highlight&quot; data-language=&quot;javascript&quot;&gt;&lt;pre class=&quot;language-javascript&quot;&gt;&lt;code class=&quot;language-javascript&quot;&gt;&lt;span class=&quot;token operator&quot;&gt;&amp;lt;&lt;/span&gt;a href&lt;span class=&quot;token operator&quot;&gt;=&lt;/span&gt;&lt;span class=&quot;token string&quot;&gt;&quot;#&quot;&lt;/span&gt;&lt;span class=&quot;token operator&quot;&gt;&gt;&lt;/span&gt;Click here&lt;span class=&quot;token operator&quot;&gt;&amp;lt;&lt;/span&gt;&lt;span class=&quot;token operator&quot;&gt;/&lt;/span&gt;a&lt;span class=&quot;token operator&quot;&gt;&gt;&lt;/span&gt;&lt;/code&gt;&lt;/pre&gt;&lt;/div&gt;
 &lt;p&gt;However, some browsers may jump (usually to the top of the page), when users click on it.&lt;/p&gt;
